move responseMap to new file
